base 1.0 preconfiged games
WASD and arrow keys for controls. designed for windows only. linux users you may need to modify the serial coms for your panels.

requires git serial and keyboard.

known issues/bugs: control script has no idea if the game is lost in snake or if a point is scored in pong. uncoment the get status at the end of the file to see what the matrix reports back.

this does not require any other API's as far as i know and should work with stock framework 16 led matrix out of the box. (as of the last prerelease batch of units).